test/requirements.txt
requests==2.32.3
boto3==1.38.19
google-cloud-speech==2.32.0
google-cloud-storage==3.1.0
vosk==0.3.45
transformers[torch]==4.52.4
torchaudio==2.6.0
azure-cognitiveservices-speech==1.44.0
jiwer==3.1.0
gigachat==0.1.39
google-genai==1.20.0
openai==1.88.0
anthropic==0.54.0
nltk==3.9.1
pymorphy3==2.0.4
rouge_score==0.1.2
bert-score==0.3.13

